[Improvement: ZSTACK-24851] zstack on kylin

+++ installation/install.sh
@@ -25,12 +25,14 @@ UBUNTU='UBUNTU'
 RHEL7='RHEL7'
 ISOFT4='ISOFT4'
 ALIOS7='AliOS7'
+KYLIN402='KYLIN4.0.2'
+DEBIAN9='DEBIAN9'
 UPGRADE='n'
 FORCE='n'
 MINI_INSTALL='n'
 SANYUAN_INSTALL='n'
 MANAGEMENT_INTERFACE=`ip route | grep default | head -n 1 | cut -d ' ' -f 5`
-SUPPORTED_OS="$CENTOS7, $UBUNTU1604, $UBUNTU1404, $ISOFT4, $RHEL7, $ALIOS7"
+SUPPORTED_OS="$CENTOS7, $UBUNTU1604, $UBUNTU1404, $ISOFT4, $RHEL7, $ALIOS7, $KYLIN402, $DEBIAN9"
 ZSTACK_INSTALL_LOG='/tmp/zstack_installation.log'
 ZSTACKCTL_INSTALL_LOG='/tmp/zstack_ctl_installation.log'
 [ -f $ZSTACK_INSTALL_LOG ] && /bin/rm -f $ZSTACK_INSTALL_LOG
@@ -194,14 +196,23 @@ vercomp () {
     done
     return 0
 }
-check_zstack_release(){
 
-    rpm -q zstack-release >/dev/null 2>&1
-    if [ $? -eq 0 ];then
-        ZSTACK_RELEASE=`rpm -qi zstack-release |awk -F ':' '/Version/{print $2}' |sed 's/ //g'`
-        source /etc/profile >/dev/null 2>&1
+check_zstack_release(){
+    if [ "$IS_YUM" = "y" ];then
+        rpm -q zstack-release >/dev/null 2>&1
+        if [ $? -eq 0 ];then
+            ZSTACK_RELEASE=`rpm -qi zstack-release |awk -F ':' '/Version/{print $2}' |sed 's/ //g'`
+            source /etc/profile >/dev/null 2>&1
+        else
+            fail2 "rpm package zstack-release is not installed, use zstack-upgrade -r/-a zstack-xxx.iso(>=3.7.0) to upgrade zstack-dvd and install zstack-release."
+        fi
     else
-        fail2 "zstack-release is not installed, use zstack-upgrade -r/-a zstack-xxx.iso(>=3.7.0) to upgrade zstack-dvd and install zstack-release."
+        dpkg -l zstack-release > /dev/null 2>&1
+        if [ $? -eq 0 ];then
+            ZSTACK_RELEASE=`awk '{print $3}' /etc/zstack-release`
+        else
+            fail2 "deb package zstack-release is not installed, use zstack_install_kylin.sh -r xxx.iso to install zstack-dvd."
+        fi
     fi
 }
 
@@ -231,6 +242,11 @@ post_scripts_to_restore_iptables_rules() {
   iptables -D INPUT `iptables -L INPUT --line-numbers | grep 'zstack allow login mysql from 127.0.0.1' | awk '{ print $1 }'`
   service iptables save
 } >>$ZSTACK_INSTALL_LOG 2>&1
+
+post_restore_source_on_debian() {
+    mv /etc/apt/sources.list.d/tmp_bak/*.list /etc/apt/sources.list.d/ 2>/dev/null
+    rm -rf /etc/apt/sources.list.d/tmp_bak/ 2>/dev/null
+}
 
 cleanup_function(){
     /bin/rm -f $UPGRADE_LOCK
@@ -606,10 +622,12 @@ cs_check_mysql_password () {
     #If user didn't assign mysql root password, then check original zstack mysql password status
     if [ 'y' != $UPGRADE ]; then
         if [ -z $ONLY_INSTALL_ZSTACK ];then
-            rpm -qa | grep mysql-community >>$ZSTACK_INSTALL_LOG 2>&1
-            if [ $? -eq 0 ];then
-                fail2 "Detect mysql-community installed, please uninstall it due to ZStack will use mariadb."
+            if [ "$IS_YUM" = "y" ];then
+                rpm -qa | grep mysql-community >>$ZSTACK_INSTALL_LOG 2>&1 && mysql_community='y'
+            else
+                dpkg -l | grep mysql-community >>$ZSTACK_INSTALL_LOG 2>&1 && mysql_community='y'
             fi
+            [ "$mysql_community" = "y" ] && fail2 "Detect mysql-community installed, please uninstall it due to ZStack will use mariadb."
         fi
 
         if [ -z $MYSQL_ROOT_PASSWORD ] && [ -z $ONLY_INSTALL_ZSTACK ]; then
@@ -675,50 +693,50 @@ cs_check_zstack_data_exist(){
 check_system(){
     echo_title "Check System"
     echo ""
-    cat /etc/*-release |egrep -i -h "centos |Red Hat Enterprise|Alibaba" >>$ZSTACK_INSTALL_LOG 2>&1
+    cat /etc/*-release |egrep -i -h "centos |Red Hat Enterprise|Alibaba|NeoKylin" >>$ZSTACK_INSTALL_LOG 2>&1
     if [ $? -eq 0 ]; then
-        grep 'CentOS release 6' /etc/system-release >>$ZSTACK_INSTALL_LOG 2>&1; IS_CENTOS_6=$?
-        grep 'CentOS Linux release 7' /etc/system-release >>$ZSTACK_INSTALL_LOG 2>&1; IS_CENTOS_7=$?
-        grep 'Red Hat Enterprise Linux Server release 7' /etc/system-release >>$ZSTACK_INSTALL_LOG 2>&1; IS_RHEL_7=$?
-        grep 'Alibaba Group Enterprise Linux' /etc/system-release >>$ZSTACK_INSTALL_LOG 2>&1; IS_ALIOS_7=$?
-        grep 'iSoft Linux release 4' /etc/system-release >>$ZSTACK_INSTALL_LOG 2>&1; IS_ISOFT_4=$?
-        if [ $IS_CENTOS_6 -eq 0 ]; then
+        grep -q 'CentOS release 6' /etc/system-release && IS_CENTOS_6='y'
+        grep -q 'CentOS Linux release 7' /etc/system-release && IS_CENTOS_7='y'
+        grep -q 'Red Hat Enterprise Linux Server release 7' /etc/system-release && IS_RHEL_7='y'
+        grep -q 'Alibaba Group Enterprise Linux' /etc/system-release && IS_ALIOS_7='y'
+        grep -q 'iSoft Linux release 4' /etc/system-release && IS_ISOFT_4='y'
+        grep -q 'NeoKylin Linux Advanced Server release V7' /etc/system-release && IS_RHEL_7='y'
+        if [ "$IS_CENTOS_6" = "y" ]; then
             OS=$CENTOS6
-        elif [ $IS_CENTOS_7 -eq 0 ]; then
+        elif [ "$IS_CENTOS_7" = "y" ]; then
             OS=$CENTOS7
             rpm -q libvirt | grep 1.1.1-29 >/dev/null 2>&1 && fail2 "Your OS is old CentOS7, as its libvirt is `rpm -q libvirt`. \
               You need to use \`yum upgrade\` to upgrade your system to latest CentOS7."
-        elif [ $IS_RHEL_7 -eq 0 ]; then
+        elif [ "$IS_RHEL_7" = "y" ]; then
             OS=$RHEL7
-        elif [ $IS_ALIOS_7 -eq 0 ]; then
+        elif [ "$IS_ALIOS_7" = "y" ]; then
             OS=$ALIOS7
-        elif [ $IS_ISOFT_4 -eq 0 ]; then
+        elif [ "$IS_ISOFT_4" = "y" ]; then
             OS=$ISOFT4
         else
             fail2 "Host OS checking failure: your system is: `cat /etc/redhat-release`, $PRODUCT_NAME management node only supports $SUPPORTED_OS currently"
         fi
     else
-        grep 'Ubuntu' /etc/issue >>$ZSTACK_INSTALL_LOG 2>&1
-        if [ $? -eq 0 ]; then
-            grep '16.04' /etc/issue >>$ZSTACK_INSTALL_LOG 2>&1
-            if [ $? -eq 0 ]; then
-                OS=$UBUNTU1604
-            else
-                grep '14.04' /etc/issue >>$ZSTACK_INSTALL_LOG 2>&1
-                if [ $? -eq 0 ]; then
-                    OS=$UBUNTU1404
-                else
-                    fail2 "Host OS checking failure: your system is: $OS, $PRODUCT_NAME management node only support $SUPPORTED_OS currently"
-                fi
-            fi
+        grep -q 'Debian GNU/Linux' /etc/issue && IS_DEBIAN='y'
+        grep -q 'Ubuntu' /etc/issue && IS_UBUNTU='y'
+        grep -q 'Kylin' /etc/issue && IS_KYLINOS='y'
+        if [ "$IS_UBUNTU" = "y" ]; then
+            grep -q '16.04' /etc/issue && OS=$UBUNTU1604
+            grep -q '14.04' /etc/issue && OS=$UBUNTU1404
+            [ "$OS" != "$UBUNTU1604" -a "$OS" != "$UBUNTU1404" ] && fail2 "Host OS checking failure: your system is: $OS, $PRODUCT_NAME management node only support $SUPPORTED_OS currently"
             . /etc/lsb-release
+        elif [ "$IS_KYLINOS" = "y" ]; then
+            grep -q '4.0.2' /etc/issue && OS=$KYLIN402 || fail2 "Host OS checking failure: your system is: `cat /etc/issue`, $PRODUCT_NAME management node only support $SUPPORTED_OS currently"
+        elif [ "$IS_DEBIAN" = "y" ]; then
+            grep -q 'Linux 9' /etc/issue && OS=$DEBIAN9 || fail2 "Host OS checking failure: your system is: `cat /etc/issue`, $PRODUCT_NAME management node only support $SUPPORTED_OS currently"
         else
             fail2 "Host OS checking failure: your system is: `cat /etc/issue`, $PRODUCT_NAME management node only support $SUPPORTED_OS currently"
         fi
+        export DEBIAN_FRONTEND=noninteractive
     fi
 
-    if [ $OS != $CENTOS7 -a $OS != $UBUNTU1404 -a $OS != $UBUNTU1604 -a $OS != $ALIOS7 ]; then
-        #only support offline installation for CentoS7.x
+    if [ $OS != $CENTOS7 -a $OS != $UBUNTU1404 -a $OS != $UBUNTU1604 -a $OS != $ALIOS7 -a $OS != $RHEL7 -a $OS != $KYLIN402 -a $OS != $DEBIAN9 ]; then
+        #only support offline installation for CentoS7.x,Kylin4.0.2
         if [ -z "$YUM_ONLINE_REPO" ]; then
             fail2 "Your system is $OS . ${PRODUCT_NAME} installer can not use '-o' or '-R' option on your system. Please remove '-o' or '-R' option and try again."
         fi
@@ -757,13 +775,14 @@ cs_create_repo(){
     echo_subtitle "Update Package Repository"
     if [ $OS = $CENTOS7 -o $OS = $CENTOS6 ]; then
         create_yum_repo
+    elif [ $OS = $KYLIN402 -o $OS = $DEBIAN9 ];then
+        create_apt_source
     fi
-
-    if [ $OS = $UBUNTU1404 -o $OS = $UBUNTU1604 ]; then
-        if [ ! -z $ZSTACK_PKG_MIRROR ]; then
-            create_apt_source_list
-        fi
-    fi
+    # if [ $OS = $UBUNTU1404 -o $OS = $UBUNTU1604 ]; then
+    #     if [ ! -z $ZSTACK_PKG_MIRROR ]; then
+    #         create_apt_source_list
+    #     fi
+    # fi
     pass
 }
 
@@ -973,7 +992,7 @@ ia_install_python_gcc_rh(){
 
 ia_install_pip(){
     echo_subtitle "Install PIP"
-    rpm -q python2-pip >/dev/null 2>&1 && return
+    which pip >/dev/null 2>&1 && which pip2 >/dev/null && return
 
     if [ ! -z $DEBUG ]; then
         easy_install -i $pypi_source_easy_install --upgrade pip
@@ -1052,6 +1071,15 @@ iu_deploy_zstack_repo() {
     chown -R zstack:zstack ${ZSTACK_HOME}/static/zstack-repo
 }
 
+iu_deploy_zstack_apt_source() {
+    echo_subtitle "Deploy apt source for ${PRODUCT_NAME}"
+
+    mkdir -p ${ZSTACK_HOME}/static/zstack-repo
+    ln -s /opt/zstack-dvd/x86_64 ${ZSTACK_HOME}/static/zstack-repo/x86_64 >/dev/null 2>&1
+    ln -s /opt/zstack-dvd/aarch64 ${ZSTACK_HOME}/static/zstack-repo/aarch64 >/dev/null 2>&1
+    chown -R zstack:zstack ${ZSTACK_HOME}/static/zstack-repo
+}
+
 unpack_zstack_into_tomcat(){
     echo_title "Install ${PRODUCT_NAME} Package"
     echo ""
@@ -1061,7 +1089,11 @@ unpack_zstack_into_tomcat(){
     fi
     show_spinner iz_unzip_tomcat
     show_spinner iz_install_zstack
-    show_spinner iu_deploy_zstack_repo
+    if [ $OS = $CENTOS7 -o $OS = $CENTOS6 -o $OS = $RHEL7 -o $OS = $ISOFT4 -o $OS = $ALIOS7 ]; then
+        show_spinner iu_deploy_zstack_repo
+    elif [ $OS = $UBUNTU1404 -o $OS = $UBUNTU1604 -o $OS = $KYLIN402 -o $OS = $DEBIAN9 ]; then
+        show_spinner iu_deploy_zstack_apt_source
+    fi
 }
 
 upgrade_zstack(){
@@ -1093,7 +1125,11 @@ upgrade_zstack(){
     show_spinner uz_stop_zstack_ui
     show_spinner uz_upgrade_zstack
     show_spinner upgrade_tomcat_security
-    show_spinner iu_deploy_zstack_repo
+    if [ $OS = $CENTOS7 -o $OS = $CENTOS6 -o $OS = $RHEL7 -o $OS = $ISOFT4 -o $OS = $ALIOS7 ]; then
+        show_spinner iu_deploy_zstack_repo
+    elif [ $OS = $UBUNTU1404 -o $OS = $UBUNTU1604 -o $OS = $KYLIN402 -o $OS = $DEBIAN9 ]; then
+        show_spinner iu_deploy_zstack_apt_source
+    fi
     show_spinner cp_third_party_tools
 
     cd /
@@ -1292,8 +1328,7 @@ install_ansible(){
     if [ $OS = $CENTOS7 -o $OS = $CENTOS6 -o $OS = $RHEL7 -o $OS = $ISOFT4 -o $OS = $ALIOS7 ]; then
         show_spinner ia_disable_selinux
         show_spinner ia_install_python_gcc_rh
-    elif [ $OS = $UBUNTU1404 -o $OS = $UBUNTU1604 ]; then
-        export DEBIAN_FRONTEND=noninteractive
+    elif [ $OS = $UBUNTU1404 -o $OS = $UBUNTU1604 -o $OS = $KYLIN402 -o $OS = $DEBIAN9 ]; then
         #if [ -z $ZSTACK_PKG_MIRROR ]; then
         #    show_spinner ia_update_apt
         #fi
@@ -1304,7 +1339,7 @@ install_ansible(){
 
 iz_install_unzip(){
     echo_subtitle "Install unzip"
-    if [ $OS = $UBUNTU1404 -o $OS = $UBUNTU1604 ]; then
+    if [ $OS = $UBUNTU1404 -o $OS = $UBUNTU1604 -o $OS = $KYLIN402 -o $OS = $DEBIAN9 ]; then
         apt-get -y install unzip >>$ZSTACK_INSTALL_LOG 2>&1
         [ $? -ne 0 ] && fail "Install unzip fail."
         pass
@@ -1421,37 +1456,41 @@ is_install_general_libs_deb(){
     fi
     openjdk=openjdk-8-jdk
 
-    if [ $OS == $UBUNTU1404 ]; then
+    if [ $OS == $UBUNTU1404 -o $OS == $KYLIN402 -o $OS = $DEBIAN9 ]; then
         #install openjdk ppa for openjdk-8
-        add-apt-repository ppa:openjdk-r/ppa -y >>$ZSTACK_INSTALL_LOG 2>&1
-        apt-get update  >>$ZSTACK_INSTALL_LOG 2>&1
-    fi
 
-    apt-get -y install \
+        # add-apt-repository ppa:openjdk-r/ppa -y >>$ZSTACK_INSTALL_LOG 2>&1
+        apt-key update >>$ZSTACK_INSTALL_LOG 2>&1
+        apt-get update >>$ZSTACK_INSTALL_LOG 2>&1
+    fi
+    apt-get -y install --allow-unauthenticated \
         $openjdk \
         bridge-utils \
         wget \
-        python-libvirt \
-        libvirt-bin \
+        curl \
         vlan \
         python-dev \
         gcc \
+        chrony \
         python-setuptools \
         >>$ZSTACK_INSTALL_LOG 2>&1
     [ $? -ne 0 ] && fail "install system lib 1 failed"
 
-    apt-get -y install --no-upgrade \
+    apt-get -y install --no-upgrade --allow-unauthenticated \
         sudo \
         >>$ZSTACK_INSTALL_LOG 2>&1
     [ $? -ne 0 ] && fail "install system lib 2 failed"
 
-    apt-get -y install \
+    apt-get -y install --allow-unauthenticated \
+        dmidecode \
         nfs-common \
         nfs-kernel-server \
         autoconf \
-        iptables-persistent \
+        netfilter-persistent \
         tar \
         gzip \
+        grafana \
+        sqlite3 \
         unzip \
         apache2 \
         sshpass \
@@ -1478,6 +1517,9 @@ is_install_general_libs_deb(){
             && ln -s /etc/init.d/netfilter-persistent /etc/init.d/iptables-persistent
         [ ! -f /etc/init.d/iptables ] && [ -f /etc/init.d/netfilter-persistent ] \
             && ln -s /etc/init.d/netfilter-persistent /etc/init.d/iptables
+        sudo /bin/systemctl daemon-reload >>$ZSTACK_INSTALL_LOG 2>&1
+        sudo /bin/systemctl enable grafana-server >>$ZSTACK_INSTALL_LOG 2>&1
+        sudo /bin/systemctl start grafana-server >>$ZSTACK_INSTALL_LOG 2>&1
     fi
     pass
 }
@@ -1878,7 +1920,6 @@ iz_unzip_tomcat(){
     if [ $? -ne 0 ];then
        fail "failed to unzip Tomcat package: $ZSTACK_INSTALL_ROOT/apache-tomcat*.zip."
     fi
-    unzip -o -d apache-tomcat/webapps/ libs/tomcat_root_app.zip >>$ZSTACK_INSTALL_LOG 2>&1
     apache_temp=`mktemp`
     apache_zip=`ls apache-tomcat*.zip`
     mv $apache_zip $apache_temp
@@ -1888,6 +1929,7 @@ iz_unzip_tomcat(){
     #delete unused web app folders, 'ROOT' should be left
     find $ZSTACK_INSTALL_ROOT/apache-tomcat/webapps -mindepth 1 -not -name 'ROOT' -delete
 
+    unzip -o -d apache-tomcat/webapps/ libs/tomcat_root_app.zip >>$ZSTACK_INSTALL_LOG 2>&1
     chmod a+x apache-tomcat/bin/*
     if [ $? -ne 0 ];then
        fail "chmod failed in: $ZSTACK_INSTALL_ROOT/apache-tomcat/bin/*."
@@ -2000,6 +2042,11 @@ setup_install_param(){
     if [ x"$SANYUAN_INSTALL" = x"y" ];then
         zstack-ctl configure identity.init.type="PRIVILEGE_ADMIN"
         zstack-ctl configure sanyuan.installed=true
+    fi
+
+    # Port 9090 is already used by system service on KylinOS(ft2000)
+    if [ $OS == $KYLIN402 ];then
+        zstack-ctl configure Prometheus.port=9080
     fi
 }
 
@@ -2329,7 +2376,7 @@ cs_setup_nfs(){
 cs_setup_nginx(){
     echo_subtitle "Configure Nginx Server"
 mkdir -p /etc/nginx/conf.d/mn_pxe/ && chmod -R 0777 /etc/nginx/conf.d/mn_pxe/
-cp -f /etc/nginx/nginx.conf /etc/nginx/nginx.conf.bck
+[ -f /etc/nginx/nginx.conf ] && cp -f /etc/nginx/nginx.conf /etc/nginx/nginx.conf.bck
 cat > /etc/nginx/nginx.conf << EOF
 user nginx;
 worker_processes auto;
@@ -2629,6 +2676,33 @@ EOF
     fi
 }
 
+create_apt_source(){
+    /bin/cp -f /etc/apt/sources.list /etc/apt/sources.list.zstack.`date +%Y-%m-%d_%H-%M-%S` >>$ZSTACK_INSTALL_LOG 2>&1
+    cat > /etc/apt/sources.list << EOF
+deb file:///opt/zstack-dvd/$BASEARCH/$ZSTACK_RELEASE/ Packages/
+EOF
+    if [ x"$UPGRADE" = x"y" ];then
+        mkdir /etc/apt/sources.list.d/tmp_bak
+        mv /etc/apt/sources.list.d/*.list /etc/apt/sources.list.d/tmp_bak 2>/dev/null
+    fi
+    #Fix Ubuntu conflicted dpkg lock issue. 
+    if [ -f /etc/init.d/unattended-upgrades ]; then
+        /etc/init.d/unattended-upgrades stop  >>$ZSTACK_INSTALL_LOG 2>&1
+        update-rc.d -f unattended-upgrades remove >>$ZSTACK_INSTALL_LOG 2>&1
+        pid=`lsof /var/lib/dpkg/lock|grep lock|awk '{print $2}'`
+        [ ! -z $pid ] && kill -9 $pid >>$ZSTACK_INSTALL_LOG 2>&1
+        which systemctl >/dev/null 2>&1
+        [ $? -eq 0 ] && systemctl stop apt-daily >>$ZSTACK_INSTALL_LOG 2>&1
+    fi
+    dpkg --configure --force-confold -a >>$ZSTACK_INSTALL_LOG 2>&1
+    [ $? -ne 0 ] && fail "execute \`dpkg --configure --force-confold -a\` failed."
+    apt-get clean >>$ZSTACK_INSTALL_LOG 2>&1
+    apt-get update -o Acquire::http::No-Cache=True --fix-missing>>$ZSTACK_INSTALL_LOG 2>&1
+    if [ $? -ne 0 ]; then
+        fail "apt-get update package failed."
+    fi
+}
+
 
 #create zstack local yum repo
 create_yum_repo(){
@@ -2831,6 +2905,15 @@ check_hybrid_arch(){
     return 1
 }
 
+create_local_source_list_files() {
+    echo "create $list_file" >> $ZSTACK_INSTALL_LOG
+    
+list_file=/etc/apt/sources.list.d/zstack-local.list
+cat > $list_file << EOF
+deb file:///opt/zstack-dvd/$BASEARCH/$ZSTACK_RELEASE/ Packages/
+EOF
+}
+
 check_sync_local_repos() {
   echo_subtitle "Check local repo version"
   [ -f ".repo_version" ] || fail2 "Cannot found current repo_version file, please make sure you have correct zstack-installer package."
@@ -2841,7 +2924,7 @@ check_sync_local_repos() {
   fi
   if [ $? -eq 0 ]; then
       return 0
-  elif [ x"$SKIP_SYNC" = x'y' ]; then
+  elif [ x"$SKIP_SYNC" = x'y' -o $OS == $KYLIN402 -o $OS = $DEBIAN9 ]; then
       echo " ... $(tput setaf 1)NOT MATCH$(tput sgr0)" | tee -a $ZSTAC_INSTALL_LOG
       echo_hints_to_upgrade_iso
   else
@@ -3118,6 +3201,10 @@ get_mn_port
 # Fix ZSTAC-22364
 pre_scripts_to_adjust_iptables_rules
 
+
+which yum >>$ZSTACK_INSTALL_LOG 2>&1 && IS_YUM='y'
+which apt >>$ZSTACK_INSTALL_LOG 2>&1 && IS_APT='y'
+
 if [ x"$ZSTACK_OFFLINE_INSTALL" = x'y' ]; then
     if [ ! -d /opt/zstack-dvd/$BASEARCH/ ]; then
         cd /opt/zstack-dvd/Packages/
@@ -3312,9 +3399,17 @@ export HISTTIMEFORMAT HISTSIZE HISTFILESIZE PROMPT_COMMAND
 EOF
 }
 
-# make sure local repo exist and dependences installed
-install_sync_repo_dependences
-create_local_repo_files
+# make sure local repo files exist
+if [ "$IS_YUM" = "y" ];then
+    # make sure local repo exist and dependences installed
+    install_sync_repo_dependences
+    create_local_repo_files
+elif [ "$IS_APT" = "y" ];then
+    create_local_source_list_files
+else 
+    fail2 "Command yum or apt not found, cannot create local repo."
+fi
+
 
 # CHECK_REPO_VERSION
 if [ x"${CHECK_REPO_VERSION}" == x"True" ]; then
@@ -3476,6 +3571,9 @@ if [ x"$UPGRADE" = x'y' ]; then
     echo_star_line
     start_zstack_tui
     post_scripts_to_restore_iptables_rules
+    if [ $OS == $KYLIN402 -o $OS = $DEBIAN9 ];then
+        post_restore_source_on_debian
+    fi
     exit 0
 fi
 
@@ -3646,3 +3744,4 @@ check_ha_need_upgrade
 echo_star_line
 start_zstack_tui
 post_scripts_to_restore_iptables_rules
+
